Fix for 32-bit (esp. win32 failing build).

+++ src/main/command_executor_portfolio.cpp
@@ -304,7 +304,7 @@ bool CommandExecutorPortfolio::doCommandSingleton(Command* cmd)
                          &d_channelsIn[0],
                          &d_smts[0]);
 
-    uint64_t threadStackSize = d_options[options::threadStackSize];
+    size_t threadStackSize = d_options[options::threadStackSize];
     threadStackSize *= 1024 * 1024;
 
     pair<int, bool> portfolioReturn =
